refactor: drop BackgroundService suffix from server class names
Rename OpcUaSubjectServerBackgroundService to OpcUaSubjectServer (the
existing internal SDK wrapper becomes OpcUaStandardServer to free the
name). Rename MqttSubjectServerBackgroundService to MqttSubjectServer.
WebSocketSubjectServer was already named without the suffix. Also
remove the obsolete src/Namotion.Interceptor.Mqtt/summary.md changelog
artifact.

Note: git's --stat output for the OPC UA pair appears as
"new file + delete + modify" rather than "renamed" because the file
names rotate (the SDK wrapper file's previous name becomes the BG
service file's new name). Content moves are clean.
